+++ app/src/main/java/com/androiddev/astrohelpme/ui/activity/AuthActivity.kt
@@ -1,14 +1,8 @@
 package com.androiddev.astrohelpme.ui.activity
 
 import android.os.Bundle
-import android.util.Log
 import androidx.appcompat.app.AppCompatActivity
-import androidx.databinding.DataBindingUtil
-import androidx.lifecycle.ViewModelProvider
 import com.androiddev.astrohelpme.R
-import com.androiddev.astrohelpme.databinding.ActivityAuthBinding
-import com.androiddev.astrohelpme.ui.viewmodel.AuthViewModel
-import com.androiddev.astrohelpme.utils.api.Resource
 import com.androiddev.astrohelpme.utils.extns.fullScreen
 import dagger.hilt.android.AndroidEntryPoint
