+++ app/src/main/java/com/androiddev/astrohelpme/ui/viewmodel/auth/AuthViewModel.kt
@@ -1,0 +1,16 @@
+package com.androiddev.astrohelpme.ui.viewmodel.auth
+
+import com.androiddev.astrohelpme.data.repository.AuthRepository
+import com.androiddev.astrohelpme.ui.viewmodel.BaseViewModel
+import dagger.hilt.android.lifecycle.HiltViewModel
+import javax.inject.Inject
+
+
+@HiltViewModel
+class AuthViewModel @Inject constructor(
+    private val repository: AuthRepository,
+) : BaseViewModel() {
+
+
+
+}
