+++ app/src/main/res/layout/fragment_login.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="viewModel"
-            type="com.androiddev.astrohelpme.ui.viewmodel.LoginViewModel" />
+            type="com.androiddev.astrohelpme.ui.viewmodel.auth.LoginViewModel" />
     </data>
 
     <FrameLayout
