+++ app/src/main/res/layout/fragment_otp_verify.xml
@@ -9,7 +9,7 @@
 
         <variable
             name="viewModel"
-            type="com.androiddev.astrohelpme.ui.viewmodel.OtpVerifyViewModel" />
+            type="com.androiddev.astrohelpme.ui.viewmodel.auth.OtpVerifyViewModel" />
     </data>
