+++ app/src/main/res/layout/fragment_signup.xml
@@ -9,7 +9,7 @@
 
         <variable
             name="viewModel"
-            type="com.androiddev.astrohelpme.ui.viewmodel.SignUpViewModel" />
+            type="com.androiddev.astrohelpme.ui.viewmodel.auth.SignUpViewModel" />
     </data>
 
     <FrameLayout
